pyproject.toml
[tool.poetry]
name = "split-fasta-by-sum-seq-len"
version = "0.1.0"
description = "Split fasta file to chunks by max sum length of sequences."
authors = ["sharvadim07 <sharvadim07@yandex.ru>"]
readme = "README.md"
packages = [{include = "split_fasta_by_sum_seq_len"}]

[tool.poetry.dependencies]
python = "^3.11"
biopython = "^1.81"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
